Add files via upload
This is a program that recognizes speech from numbers 1 to  9 and displays them in the seven segment display using arduino. 
- steps to follow
1. use the .png image file to draw the circuit of the arduino
1. Upload the .ino file using Arduino ide to your Arduino uno.
2. Open the .py file and run the code and say the number you want to say.

Note:  you might want to install "pip install pyserial " and "pip install SpeechRecognition" on your CMD. Also the speech_recognition uses a Google AI so you need to be connected to an internet.